add beers routes
