+++ api/users/[id].js
@@ -2,7 +2,7 @@ import bcrypt from 'bcrypt'
 import { ObjectId } from 'mongodb'
 
 import getClient from '../../mongodb.js'
-import auth from './_auth.js'
+import auth from '../../middleware/_auth.js'
 
 export default async function(req, res) {
     if (req.method === 'OPTIONS') {
